Support get topic applied policy for MaxUnackedMessagesOnConsumer

+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/admin/AdminApiMaxUnackedMessages.java
@@ -52,9 +52,9 @@ public class AdminApiMaxUnackedMessages extends MockedPulsarServiceBaseTest {
     public void testMaxUnackedMessagesOnConsumers() throws Exception {
         admin.namespaces().createNamespace("max-unacked-messages/default-on-consumers");
         String namespace = "max-unacked-messages/default-on-consumers";
-        assertEquals(50000, admin.namespaces().getMaxUnackedMessagesPerConsumer(namespace));
+        assertEquals(50000, admin.namespaces().getMaxUnackedMessagesPerConsumer(namespace).intValue());
         admin.namespaces().setMaxUnackedMessagesPerConsumer(namespace, 2*50000);
-        assertEquals(2*50000, admin.namespaces().getMaxUnackedMessagesPerConsumer(namespace));
+        assertEquals(2*50000, admin.namespaces().getMaxUnackedMessagesPerConsumer(namespace).intValue());
     }
 
     @Test
